Seperate Mainpage and user response page(TellYourStoryPage)
